+++ resources/views/dashboard/dashcompany.blade.php
@@ -14,6 +14,7 @@
     @section ('seccion')
         <h1>Dashboard Companies</h1>
         <a href="{{ route('dashboard')}}" class="btn btn-primary">Go back</a>
+        <a href="{{route('CompanyPdf')}}" class="btn btn-secondary">Download PDF</a>
         <table class="table" style="width:80vw; margin-left:-50px;">
             <thead>
                 <tr>
@@ -37,7 +38,7 @@
                         <td>{{$company->companyWeb}}</td>
                         <td>{{$company->companyLogo}}</td>
                         <td><a href="/admin/companymod?companyId={{$company->companyId}}" class="btn btn-warning">Modify</a></td>
-                        <td><a href="companyDelete?companyId={{$company->companyId}}" class="btn btn-danger">Delete</a></td>
+                        <td><a onclick="return confirm('Seguro que quiere borrar?')" href="companyDelete?companyId={{$company->companyId}}" class="btn btn-danger">Delete</a></td>
                     </tr>
                 @endforeach
             </tbody>
